pyproject.toml
[tool.poetry]
name = "secretarai"
version = "0.1.0"
description = "AI Personal Assistant - Telegram-based with LLM-agnostic architecture"
authors = ["Luca"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Telegram Bot
python-telegram-bot = "^21.0"
# LLM Clients (Native)
google-genai = "^1.0.0"
anthropic = "^0.40.0"
# Google Workspace
google-api-python-client = "^2.150.0"
google-auth-oauthlib = "^1.2.0"
# Voice Transcription
openai = "^1.50.0"
# Data & Config
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
python-dotenv = "^1.0.1"
# Database
sqlalchemy = "^2.0.35"
aiosqlite = "^0.20.0"
# HTTP & Async
httpx = "^0.27.0"
tenacity = "^9.0.0"
# Geolocation
geopy = "^2.4.1"
# PC Automation (optional)
pyautogui = "^0.9.54"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
